Change directory structure
I have changed the directory structure
which allows a "binning" effect. However
this means I have to remove the index.md
file for resources, which is a little awkward.

I'm pretty sure I'll have to customize the hugo
partials on the theme to allow for this to work
with an existing index.md file